Add morning-affirmations route with model
- add morning-affirmations route
- add morning-affirmation model that fetches all morning_affirmations
from the server
- morning-affirmations template successfully renders the 'completed'
attribute of each morningAffirmation in the UI

The app still breaks when I try to include relationships in the model.
If I uncomment the 'belongs-to' relationshop that morningAffirmations
have with an affirmation and a morning, the app errors out and nothing
renders. This may require a custom serializer, or a differently
formatted response from the API.